+++ api/src/services/submission/support/problem_support_trait.rs
@@ -1,7 +1,7 @@
 use crate::models::submission::submission_test::SubmissionTest;
 
 pub struct SubmissionGenerationPayload {
-    pub seed: i32,
+    pub seed: i64,
     pub sample_index: Option<i32>,
 }
 
@@ -10,7 +10,7 @@ pub struct SubmissionGenerationResult {
 }
 
 pub struct SubmissionTestGenerationPayload {
-    pub seed: i32,
+    pub seed: i64,
     pub test_index: usize,
     pub sample_index: Option<i32>,
 }
